requirements.txt
# Core dependencies
python-dotenv==1.0.0

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
selenium==4.16.0
aiohttp==3.9.3
async-timeout==4.0.3

# Data storage
sqlalchemy==2.0.25

# NLP and embeddings
sentence-transformers==2.3.1
transformers==4.37.2
torch==2.2.0
numpy==1.26.3
pandas==2.2.0

# Search and retrieval
faiss-cpu==1.7.4
annoy==1.17.3

# AI/LLM integration
openai==1.10.0
langchain==0.1.4
langchain-community==0.0.16

# UI
streamlit==1.31.0
plotly==5.18.0

# Utilities
python-dateutil==2.8.2
tqdm==4.66.1
pyyaml==6.0.1
validators==0.22.0

# Development
pytest==8.0.0
black==24.1.1
flake8==7.0.0

# Logging
loguru==0.7.2
rich==13.7.0

